Adjust breakpoints and associated layout
Breakpoints were adjusted to accommodate all screen sizes regardless
of device type.

The text on the main section gets squeezed at a specific breakpoint
so the flex direction was changed.

A wrap behavior was also added on the navigation list so that links
with longer text will not cause an overflow.